Refactor toasts
