+++ src/components/Board/Board.js
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react';
 
 import useBoardContent from '@/features/structure/useBoardContent';
 import {useSaveState, useHydrateState} from '@/features/localStorage';
-import {useInfoToast} from '@/components/InfoToast';
+import {useInfoToast} from '@/components/Toasts/InfoToast';
 import {Row, Col} from '@/ui/Grid';
 import Column from '@/components/Column';
