+++ src/components/Node/Node.js
@@ -7,7 +7,7 @@ import {useNodeData, useChildrenCompletion} from '@/features/structure';
 import {Button} from '@/ui/Button';
 import {FooterSeparator} from '@/ui/FooterSeparator';
 import {PieChart} from '@/components/PieChart';
-import {useInfoToast} from '@/components/InfoToast';
+import {useInfoToast} from '@/components/Toasts/InfoToast';
 import {useDeleteNodeModal} from '@/components/Modals/ConfirmationDeleteNode';
 import {useEditNodeModal} from '@/components/Modals/EditNode';
